+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 - Rejects with `FetchError` instead of `HttpError`.
 - Adds the `request` property (an instance of [`Request`][]) with details of the
   request to the error.
+- Rejects with `FetchError` also for network errors.
 
 [`Request`]: https://developer.mozilla.org/en-US/docs/Web/API/Request